Add genaric functions file with search and fs.read functions Relates #33
